Refactoring: change directory name
