+++ include/tweedledum/algorithms/synthesis/xag/CollapsedXAG.h
